soil_analysis/requirements.txt
# Core web framework
fastapi>=0.100.0
uvicorn[standard]>=0.20.0

# Data processing and ML
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.2.0
joblib>=1.2.0

# Data validation
pydantic>=2.0.0

# Database (optional for future use)
sqlalchemy>=2.0.0

# Configuration
python-dotenv>=1.0.0

# Logging
loguru>=0.7.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.20.0
httpx>=0.24.0

# Development
black>=23.0.0
flake8>=6.0.0
